Day_4/ai_assisstant_project/backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-groq
